+++ djangojs/templatetags/js.py
@@ -157,3 +157,13 @@ def django_js(context, jquery=True, i18n=True, csrf=True):
             'csrf': _boolean(csrf),
         }
     }
+
+@register.inclusion_tag('djangojs/django_js_init.html', takes_context=True)
+def django_js_amd(context, i18n=True, csrf=True):
+    '''Include Django.js javascript library in the page'''
+    return {
+        'js': {
+            'i18n': _boolean(i18n),
+            'csrf': _boolean(csrf),
+        }
+    }