+++ djangojs/__init__.py
@@ -3,7 +3,7 @@
 Django.js provide better integration of javascript into Django.
 '''
 
-__version__ = '0.7.dev'
+__version__ = '0.7.0'
 __description__ = "Django JS Tools"
 
 #: Packaged jQuery version