+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Current
 -------
 
 - Added RequireJS/AMD helpers and documentation
+- Added Python 3 support (thanks to Travis Jensen)
 
 
 0.6.5 (2013-03-13)